+++ src/server/server.c
@@ -72,10 +72,10 @@ void *Connection(void *argv) {
                 strcat(helloMsg, user->name);
                 strcat(helloMsg, "!\n");
 
-                long size = strlen(helloMsg);
+                int size = strlen(helloMsg);
                 long encMsg[size];
                 encrypt(helloMsg, encMsg, size, user->e, user->n);
-                write(fd, &size, sizeof(long));
+                write(fd, &size, sizeof(int));
                 write(fd, encMsg, size * sizeof(long));
 
                 printUserLogMsg(fd, user->name, "joined chat");
@@ -91,10 +91,10 @@ void *Connection(void *argv) {
 
             for (int i = 0; i < count; ++i) { // Проходимся по массиву сокетов
                 if(usersArr[i].fd != fd && nicknames[i] != NULL) {
-                    long size = strlen(newBuffer);
+                    int size = strlen(newBuffer);
                     long encMsg[size];
                     encrypt(newBuffer, encMsg, size, usersArr[i].e, usersArr[i].n);
-                    write(usersArr[i].fd , &size, sizeof(long));
+                    write(usersArr[i].fd , &size, sizeof(int));
                     write(usersArr[i].fd , encMsg, size * sizeof(long)); // Отправляем сообщение всем кроме нас
                 }
             }
@@ -115,10 +115,10 @@ void *Connection(void *argv) {
             if (user->msgCount != 0) {
                 for (int i = 0; i < count; ++i) { // Проходимся по массиву сокетов
                     if(usersArr[i].fd != fd && nicknames[i] != NULL) {
-                        long size = strlen(buffer);
+                        int size = strlen(buffer);
                         long encMsg[size];
                         encrypt(buffer, encMsg, size, usersArr[i].e, usersArr[i].n);
-                        write(usersArr[i].fd , &size, sizeof(long));
+                        write(usersArr[i].fd , &size, sizeof(int));
                         write(usersArr[i].fd , encMsg, size * sizeof(long)); // Отправляем сообщение всем кроме нас
                     }
                 }
