+++ src/server/serverHandlers.c
@@ -33,8 +33,8 @@ void serverKeyHandler(struct users *user, struct keys *key, int fd) {
 }
 
 char *serverMsgHandler(int fd, int *valread, int d, int n) {
-    long size;
-    read(fd, &size, sizeof(long));
+    int size;
+    read(fd, &size, sizeof(int));
 
     long encMsg[size];
     char *buffer = (char *) malloc(size + 1);
